Prepare lib
