+++ src/cuzk/test/decompose_shader.rs
@@ -6,8 +6,9 @@ use wgpu::CommandEncoderDescriptor;
 use crate::cuzk::{
     gpu::{
         get_adapter, get_device, read_from_gpu_test,
-    }, lib::{points_to_bytes, scalars_to_bytes}, msm::{convert_point_coords_and_decompose_shaders, P, PARAMS, WORD_SIZE}, shader_manager::ShaderManager, utils::{bytes_to_field, debug, to_biguint_le}
+    }, msm::{convert_point_coords_and_decompose_shaders, P, PARAMS, WORD_SIZE}, shader_manager::ShaderManager, utils::{bytes_to_field, debug, to_biguint_le}
 };
+use crate::{points_to_bytes, scalars_to_bytes};
 
 pub async fn decompose_shader<C: CurveAffine>(points: &[C], scalars: &[C::Scalar]) -> (Vec<C>, Vec<u8>) {
     let input_size = scalars.len();
@@ -138,7 +139,7 @@ pub async fn run_webgpu_decompose_async<C: CurveAffine>(points: &[C], scalars: &
 
 #[cfg(test)]
 mod tests {
-    use crate::cuzk::lib::{sample_points, sample_scalars};
+    use crate::{sample_points, sample_scalars};
 
     use super::*;
     use halo2curves::bn256::{Fr, G1Affine};
